Refactoring reducer functions.
