+++ frontend/src/services/constants.service.ts
